Move LowercaseString near its only consumer.
